pyproject.toml
[project]
name = "reseach-paper-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.3",
    "pypdf>=3.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
    "scikit-learn>=1.3.0",
    "graphviz>=0.20.0",
]
